Organize things as platformio project.
PlatformIO is an ecosystem consisting of a buildsystem, a dependency-
management for libraries, an ide and a lot more. It makes developing
embedded projects very easy because it integrates all the tools and
frameworks that are needed. A user just needs to install platformio and
call
  $ platformio run
to build the project. All the necessary frameworks, tools and libraries
will be downloaded and correctly used.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>