Integrate WifiManager for interactive configuration.
This integrates the WifiManager that opens a softAP and spawns a
captive portal where the user can configure the WiFi settings
interactive.

If there is an SSID/Password set in the config.h, then the WifiManager
will NOT used and instead a static connection to the configured network
will get established.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>