Logging: Add websockets log target.
This adds a WebSocket log target. The log is pushed to a websocket
server where you can get it with a websockets client. To enable the
websockets-log you need to configure it. As long we don't have an
actual frontend this can be achieved with curl:
  curl -d '{"webLogLevel": "debug"}' -XPUT  http://<ip>/config

The output can the be retrieved via websockets from ws://<ip>:81/. As
an example one can use:
  https://github.com/websocket-client/websocket-client

This has a tool wsdump.py. Just point it to the endpoint:
  wsdump.py ws://<ip>:81

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>