src/MQTT433gateway.cpp: Use WiFiManager
The ESP8266 saves the wifi configuration at the end of the flash and
can use this information also after re-flashing. Since we switch to
WiFiManager, we do not anymore need to take care of this information
in our program code. Thus, the user can flash an update of
MQTT433gateway and still connect to wifi without seeing the
WiFiManager.