Wifi: change wifi handling.
This adapts PR #22.

The wifimanager is only locally instantiated and all ressources are
freed after the configuration. It also uses the device-name/config-pw
from the settings for the AP-ssid/AP-password.

The device will be rebooted if we have a new config. This should adress
the concern that the heap gets fragmented by the config portal.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>